Don’t add data to xhr if params are empty object.
This adds a new isEmptyObject util that’s used in makeXhr to not add data to the XHR request if the params are an empty object.  This fixes an issue with the XHR requests in done-ssr.

Also moves the makeXhr function up next to the rest function, since they’re usually debugged together.